refactor(app): clean up and delete ProtocolSetup and RunSetupCard
closes RAUT-286
