+++ app/src/organisms/Devices/HeaterShakerWizard/__tests__/PowerOn.test.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { renderWithProviders } from '@opentrons/components'
 import { i18n } from '../../../../i18n'
-import { PowerOn } from '../PowerOn'
 import { mockHeaterShaker } from '../../../../redux/modules/__fixtures__'
+import { PowerOn } from '../PowerOn'
 
 const render = (props: React.ComponentProps<typeof PowerOn>) => {
   return renderWithProviders(<PowerOn {...props} />, {
