+++ app/src/organisms/Devices/ProtocolRun/SetupModules/SetupModulesList.tsx
@@ -26,8 +26,8 @@ import {
 import { Banner } from '../../../../atoms/Banner'
 import { StyledText } from '../../../../atoms/text'
 import { StatusLabel } from '../../../../atoms/StatusLabel'
-import { UnMatchedModuleWarning } from '../../../ProtocolSetup/RunSetupCard/ModuleSetup/UnMatchedModuleWarning'
-import { MultipleModulesModal } from '../../../ProtocolSetup/RunSetupCard/ModuleSetup/MultipleModulesModal'
+import { UnMatchedModuleWarning } from './UnMatchedModuleWarning'
+import { MultipleModulesModal } from './MultipleModulesModal'
 import {
   ModuleRenderInfoForProtocol,
   useIsOT3,
