+++ app/src/organisms/Devices/__tests__/ModuleInfo.test.tsx
@@ -4,10 +4,10 @@ import { when, resetAllWhenMocks } from 'jest-when'
 import { ModuleModel, ModuleType } from '@opentrons/shared-data'
 import { ModuleInfo } from '../ModuleInfo'
 import { renderWithProviders } from '@opentrons/components'
-import { i18n } from '../../../../../i18n'
-import { useRunHasStarted } from '../../../../../organisms/Devices/hooks'
+import { i18n } from '../../../i18n'
+import { useRunHasStarted } from '../hooks'
 
-jest.mock('../../../../../organisms/Devices/hooks')
+jest.mock('../hooks')
 
 const mockUseRunHasStarted = useRunHasStarted as jest.MockedFunction<
   typeof useRunHasStarted
