+++ app/src/organisms/Devices/hooks/index.ts
@@ -9,6 +9,7 @@ export * from './useIsRobotBusy'
 export * from './useIsRobotViewable'
 export * from './useLabwareRenderInfoForRunById'
 export * from './useLights'
+export * from './useLPCSuccessToast'
 export * from './useModuleRenderInfoForProtocolById'
 export * from './usePipetteOffsetCalibrations'
 export * from './usePipetteOffsetCalibration'
@@ -22,6 +23,7 @@ export * from './useStoredProtocolAnalysis'
 export * from './useTipLengthCalibrations'
 export * from './useUnmatchedModulesForProtocol'
 export * from './useProtocolAnalysisErrors'
+export * from './useProtocolMetadata'
 export * from './useRunStartedOrLegacySessionInProgress'
 export * from './useProtocolRunAnalyticsData'
 export * from './useRobotAnalyticsData'
