+++ app/src/organisms/LabwarePositionCheck/DeprecatedComponents/DeprecatedSummaryScreen.tsx
@@ -16,7 +16,7 @@ import {
 import { useCreateLabwareOffsetMutation } from '@opentrons/react-api-client'
 import { useTrackEvent } from '../../../redux/analytics'
 import { useCurrentRunId } from '../../ProtocolUpload/hooks'
-import { useLPCSuccessToast } from '../../ProtocolSetup/hooks'
+import { useLPCSuccessToast } from '../../Devices/hooks/useLPCSuccessToast'
 import { DeprecatedDeckMap } from './DeprecatedDeckMap'
 import { DeprecatedSectionList } from './DeprecatedSectionList'
 import { DeprecatedLabwareOffsetsSummary } from './DeprecatedLabwareOffsetsSummary'
