--- app/src/organisms/ProtocolSetup/styles.css
@@ -1,23 +1,0 @@
-@import '@opentrons/components';
-
-.deck_map {
-  margin: 0 5%;
-}
-
-.modal {
-  display: flex;
-  align-items: flex-start;
-  z-index: 100;
-}
-
-.modal_contents {
-  max-width: 100%;
-  max-height: 100%;
-  overflow-y: auto;
-}
-
-.sticky_alert {
-  position: sticky;
-  top: 3rem;
-  z-index: 100;
-}
